graphql, core: Reorganize GraphQL crates in preparation of susbcriptions
